Update nhỏ cho delete manager

+++ src/gui/frmCreateNewManager.java
@@ -708,6 +708,10 @@ public class frmCreateNewManager extends javax.swing.JInternalFrame {
 
     private void btnDeleteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnDeleteActionPerformed
         // TODO add your handling code here:
+        if(txtName.getText().trim().length()==0){
+            JOptionPane.showMessageDialog(this, "Please chose one row from table");
+            return;
+        }
         int i = JOptionPane.showConfirmDialog(this, "Do you want to delete this Manager.?");
         if (i == JOptionPane.YES_OPTION) {
             interact.User.deleteUser(txtID.getText());
